Modernize Ars Component
This commit updates the Ars component to use React.Component, it also
adds Flow definitions and eliminates the need for the Sync mixin for
this component.

In the process of doing this, I hit a few snags using `mixed` as the
type for record ids, which turned into a bit of a large Flow type
update.