Modernize Picker, eliminate Sync mixin
This commit updates the Picker component to use the most recent React
API. It also brings Flow coverage to that component, and removes the
Sync mixin in favor of a render-prop component.

In the process of doing this, I hit a snag where the Error component
was competing with the Error exception namespace. To work around this,
I've renamed the Error component to ErrorMessage.