Add pagination
This commit adds pagination to ArsArsenal. It breaks up the data
fetcher and URL building options to make URL building clearer and
adds logic ot watch the last item in a list to trigger a new page
of data.

I need some way of identicating that pagination is done. In a
follow-up commit, I'd like to add a new option for "Pagination is
done". It would be nice to reduce configuration of this to be bear
minimum.