Refresh scroll monitor when search or sort changes
Before this commit, searching only checked scroll position when the
container changed. This was an optimization I added to avoid
constantly checking scroll, possibly triggering multiple pages of
pagination at once.

To fix this, I added a key to the scroll monitor so that it totally
refreshes whenever the sorting order or search changes.